Refactored control type masking/unmasking to IMauiElementEventBinder
